requirements.txt
flask>=3.0.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
ollama>=0.1.0
pdfplumber>=0.9.0
python-docx>=0.8.11
markdown>=3.4.0
beautifulsoup4>=4.11.0
python-pptx>=0.6.21
python-dotenv>=1.0.0
numpy>=1.21.0 